bugprone linting changes

+++ libff/common/double.cpp
@@ -150,12 +150,12 @@ namespace libff {
 
   libff::bigint<1> Double::as_bigint() const
   {
-    return libff::bigint<1>(val.real());
+    return libff::bigint<1>((unsigned long) val.real());
   }
 
   unsigned long Double::as_ulong() const
   {
-    return round(val.real());
+    return round((unsigned long) val.real());
   }
 
   Double Double::squared() const
